+++ platform/plugins/marketplace/database/migrations/2024_09_17_125408_add_square_logo_to_stores_table.php
@@ -7,15 +7,21 @@ use Illuminate\Support\Facades\Schema;
 return new class () extends Migration {
     public function up(): void
     {
-        Schema::table('mp_stores', function (Blueprint $table) {
+        if (Schema::hasColumn('mp_stores', 'logo_square')) {
+            return;
+        }
+
+        Schema::table('mp_stores', function (Blueprint $table): void {
             $table->string('logo_square', 255)->nullable()->after('logo');
         });
     }
 
     public function down(): void
     {
-        Schema::table('mp_stores', function (Blueprint $table) {
-            $table->dropColumn('logo_square');
-        });
+        if (Schema::hasColumn('mp_stores', 'logo_square')) {
+            Schema::table('mp_stores', function (Blueprint $table): void {
+                $table->dropColumn('logo_square');
+            });
+        }
     }
 };
